#51 Change user/login -> login
